requirements.txt
python-telegram-bot==13.5
psutil
speedtest-cli
schedule
requests